rename plexscoring-> scoring
