+++ alphadia/workflow/peptidecentric/requantification_handler.py
@@ -6,7 +6,7 @@ from alphabase.spectral_library.base import SpecLibBase
 from alphadia.raw_data import DiaData
 from alphadia.reporting.reporting import Pipeline
 from alphadia.search.scoring.config import CandidateConfig
-from alphadia.search.scoring.plexscoring import CandidateScoring
+from alphadia.search.scoring.scoring import CandidateScoring
 from alphadia.search.scoring.utils import (
     candidate_features_to_candidates,
     multiplex_candidates,
